+++ plugins/content-manager/src/main/java/com/wazuh/contentmanager/util/Unzip.java
@@ -1,4 +1,4 @@
-package com.wazuh.contentmanager.tools;
+package com.wazuh.contentmanager.util;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -10,24 +10,24 @@ public class Unzip {
 
     private static final Logger log = LogManager.getLogger(Unzip.class);
     private static final byte[] BUFFER = new byte[1024];
-    private static final String ZIP_MIME_TYPE = "application/zip";
 
     /*
      * Unzips a ZIP file's content in a specified folder
      * @param zipFilePath Origin ZIP file path following the format: "path/file.zip"
      * @param destDirectory Unzipped files destiny path following the format: "path/"
      */
-    public void unzip(String zipFilePath, String destDirectory) {
+    public void unzip(String zipFilePath, String destDirectory) throws FileNotFoundException {
         File zipFile = new File(zipFilePath);
         if (!zipFile.exists() || !zipFile.isFile()) {
-            log.error("Error, ZIP file does not exist or is invalid: {}", zipFilePath); return;
+            log.error("Error, ZIP file does not exist or is invalid: {}", zipFilePath);
+            throw new FileNotFoundException("Error, ZIP file does not exist or is invalid: " + zipFilePath);
         }
 
         try (ZipInputStream zipIn = new ZipInputStream(new FileInputStream(zipFilePath))) {
             ZipEntry entry;
 
             while ((entry = zipIn.getNextEntry()) != null) {
-                File filePath = new File(destDirectory, entry.getName()); // Se actualiza en cada iteración
+                File filePath = new File(destDirectory, entry.getName());
 
                 try (BufferedOutputStream bos = new BufferedOutputStream(new FileOutputStream(filePath))) {
                     int len;
@@ -42,13 +42,5 @@ public class Unzip {
         } catch (IOException e) {
             log.error("Error unzipping the file due to {}", e.getMessage());
         }
-    }
-
-    public static void main(String[] args) {
-        String filePath = "temp/prueba.zip";
-        String destDirectory = "temp/";
-
-        Unzip unzipper = new Unzip();
-        unzipper.unzip(filePath, destDirectory);
     }
 }