+++ CHANGELOG.md
@@ -7,7 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## [Unreleased 5.0.x]
 
 ### Added
-- New tools directory with unzip tool.
+- New tools directory with unzip tool [(#319)](https://github.com/wazuh/wazuh-indexer-plugins/pull/319)
 
 ### Dependencies